Fixes setValue(null)-calls-delete() bug.
git-svn-id: https://svn.zib.de/opus4dev/framework/trunk@2519 0a49f362-4213-0410-86e3-e96cb6c4fbae

+++ library/Opus/Model/Field.php
@@ -305,7 +305,7 @@ class Opus_Model_Field
                     if ($submodel instanceof Opus_Model_Dependent_Abstract) $submodel->delete();
                 }
             } else {
-                $this->_value->delete();
+                if ($this->_value instanceof Opus_Model_Dependent_Abstract) $this->_value->delete();
             }
         }
